Initial commit
This is a work-in-progress. Currently, an Arduino Uno with built-in LED
light on pin 13 will work. Run the desktop app and you can send commands
to turn on and off the light. The Arduino will return over serial the
status of the light. Next step is to make the light on/off control a
relay.